laskarit/viikko2/Unicafe/src/main/java/com/mycompany/unicafe/Maksukortti.java

package com.mycompany.unicafe;

public class Maksukortti {
 
    private int saldo;
    private final double EDULLINEN =   2.6;
    private final double MAUKAS = 4.0;
 
    public Maksukortti(int saldo) {
        this.saldo = saldo;
    }
 
    public int saldo() {
        return saldo;
    }
 
    public void lataaRahaa(int lisays) {
        this.saldo += lisays;
    }
 
    public boolean otaRahaa(int maara) {
        if (this.saldo < maara) {
            return false;
        }
 
        this.saldo = this.saldo - maara;
        return true;
    }
 

    @Override
    public String toString() {
        int euroa = saldo/100;
        int senttia = saldo%100;
        return "saldo: "+euroa+"."+senttia;
    } 
    
}
